requirements.txt
google-genai
Pillow
python-dotenv
google-cloud-aiplatform